backend/requirements.txt
# Backend runtime is validated on Python 3.11-3.13.
# Python 3.14 currently triggers source builds for some pinned deps.
fastapi==0.115.0
uvicorn==0.32.0
supabase==2.13.0
openai==1.30.1
httpx==0.27.2
pypdf==5.1.0
langchain-text-splitters==0.3.2
python-multipart==0.0.17
python-dotenv==1.0.1
openpyxl==3.1.5
pandas==2.2.3
psycopg2-binary==2.9.10
pgvector==0.3.6
pydantic-settings==2.2.1
reportlab==4.2.0
requests==2.31.0
python-docx==1.1.0
numpy==1.26.4
stripe==5.4.0
